requirements.txt
asgiref==3.5.0
coverage==6.3.2
dj-database-url==0.5.0
Django==3.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
